updated files with exception handling and added service layer

+++ SpringBootRest/src/main/java/com/webonise/springBoot/dao/StudentRepository.java
@@ -3,6 +3,5 @@ package com.webonise.springBoot.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.webonise.springBoot.model.Student;
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {
-
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
